Renamed unit tests
